Move `utils/i18n` to `utils/text`
It doesn't really include translations stuff anymore. `gettext` usage not removed though!
