+++ app/src/forms/auth/ResetPassword.js
@@ -8,7 +8,7 @@ import is from 'is_js';
 
 import { urlResolve } from 'configuration/routes';
 import FormField from 'forms/fields/FormField';
-import { gettext, pgettext } from 'utils/i18n';
+import { gettext, pgettext } from 'utils/text';
 import { getFormPropTypes } from 'utils/types';
