+++ app/src/views/auth/Login.js
@@ -7,7 +7,7 @@ import AuthLayout from 'components/layouts/AuthLayout';
 import LoginForm from 'forms/auth/Login';
 import withView from 'decorators/withView';
 import { obtainToken } from 'sagas/auth/obtainTokenSaga';
-import { pgettext } from 'utils/i18n';
+import { pgettext } from 'utils/text';
 
 
 const Login = ({ onLogin }) => (
